+++ package.js
@@ -32,6 +32,7 @@ function configure( api ){
     api.use( 'pwix:i18n@1.5.7' );
     api.use( 'pwix:jstree@1.0.6' );
     api.use( 'pwix:modal@1.10.0 || 2.0.0' );
+    api.use( 'pwix:ui-bootstrap5@2.0.0' );
     api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.addFiles( 'src/client/components/prEdit/prEdit.js', 'client' );
     api.addFiles( 'src/client/components/prEditPanel/prEditPanel.js', 'client' );